+++ src/elite/test/EliteSharpLib.Tests/ShipBaseTests.cs
@@ -140,8 +140,8 @@ public class ShipBaseTests
             vec.Z = 1;
         }
 
-        float x = ((vec.X * 256 / vec.Z) + (draw.Centre.X / 2)) * draw.Scale;
-        float y = ((-vec.Y * 256 / vec.Z) + (draw.Centre.Y / 2)) * draw.Scale;
+        float x = draw.Centre.X + (vec.X * draw.Focus / vec.Z);
+        float y = draw.Centre.Y - (vec.Y * draw.Focus / vec.Z);
         return new(x, y);
     }
 
@@ -155,8 +155,8 @@ public class ShipBaseTests
             vec.Z = 1;
         }
 
-        float x = ((vec.X * 256 / vec.Z) + (draw.Centre.X / 2)) * draw.Scale;
-        float y = ((-vec.Y * 256 / vec.Z) + (draw.Centre.Y / 2)) * draw.Scale;
+        float x = draw.Centre.X + (vec.X * draw.Focus / vec.Z);
+        float y = draw.Centre.Y - (vec.Y * draw.Focus / vec.Z);
         return new(x, y);
     }